chore: error handling
